updating utils and checking weight normalization
